Removed the NotSupportedException from the DisplayName setter.
For some reason the DockLayoutManager wanted to set this property
numerous times.
I don't know why, but I just removed the exception and wrote a comment
instead.
But now if I think of it, maybe I can change the Binding in the
Dashboard so we only have a one-way binding. That might help.